Move parameter-type utils to metabase-lib
