+++ frontend/src/metabase/parameters/utils/linked-filters.js
@@ -2,7 +2,7 @@ import {
   TYPE_SUPPORTS_LINKED_FILTERS,
   FIELD_FILTER_PARAMETER_TYPES,
 } from "metabase/parameters/constants";
-import { getParameterType } from "metabase/parameters/utils/parameter-type";
+import { getParameterType } from "metabase-lib/lib/parameters/utils/parameter-type";
 
 export function canUseLinkedFilters(parameter) {
   const type = getParameterType(parameter);
